Typescript arragements, new android utility class, android modules revised shared utility.
