+++ android/src/main/java/com/visioncamerabarcodesscanner/VisionCameraBarcodesScannerPackage.kt
@@ -4,8 +4,8 @@ import com.facebook.react.ReactPackage
 import com.facebook.react.bridge.ReactApplicationContext
 import com.facebook.react.uimanager.ViewManager
 import com.facebook.react.bridge.NativeModule
-import com.mrousavy.camera.frameprocessors.FrameProcessorPluginRegistry
 
+import com.mrousavy.camera.frameprocessors.FrameProcessorPluginRegistry
 
 class VisionCameraBarcodesScannerPackage : ReactPackage {
     companion object {
