+++ src/Camera.tsx
@@ -1,33 +1,35 @@
 import React, { forwardRef, useMemo } from 'react';
 import {
-  Camera as VisionCamera,
+  Camera as CameraVision,
   useFrameProcessor,
 } from 'react-native-vision-camera';
+import { useRunOnJS } from 'react-native-worklets-core';
 import { createBarcodeScannerPlugin } from './scanBarcodes';
 import type {
-  ReadonlyFrameProcessor,
-  Frame,
-  CameraTypes,
-  ForwardedRef,
-  Barcode,
-  BarcodeScannerPlugin,
+  ScannerPlugin,
   ScannerOptions,
+  Barcode,
+  CameraProps,
+  Frame,
+  ForwardedRef,
+  ReadonlyFrameProcessor,
 } from './types';
-import { useRunOnJS } from 'react-native-worklets-core';
 
 export const Camera = forwardRef(function Camera(
-  props: CameraTypes,
+  props: CameraProps,
   ref: ForwardedRef<any>
 ) {
   const { device, callback, options, ...p } = props;
-  // @ts-ignore
+
   const { scanBarcodes } = useBarcodeScanner(options);
+
   const useWorklets = useRunOnJS(
     (data: Barcode[]): void => {
       callback(data);
     },
     [options]
   );
+
   const frameProcessor: ReadonlyFrameProcessor = useFrameProcessor(
     (frame: Frame) => {
       'worklet';
@@ -37,10 +39,11 @@ export const Camera = forwardRef(function Camera(
     },
     []
   );
+
   return (
     <>
       {!!device && (
-        <VisionCamera
+        <CameraVision
           pixelFormat="yuv"
           ref={ref}
           frameProcessor={frameProcessor}
@@ -52,8 +55,6 @@ export const Camera = forwardRef(function Camera(
   );
 });
 
-export function useBarcodeScanner(
-  options?: ScannerOptions
-): BarcodeScannerPlugin {
+export function useBarcodeScanner(options?: ScannerOptions): ScannerPlugin {
   return useMemo(() => createBarcodeScannerPlugin(options), [options]);
 }
