+++ src/index.ts
@@ -1,3 +1,9 @@
 export { Camera, useBarcodeScanner } from './Camera';
-
 export { ImageScanner } from './ImageScanner';
+export type {
+  CameraProps,
+  ScannerOptions,
+  BarcodeType,
+  Barcode,
+  Ratio,
+} from './types';
