+++ src/scanBarcodes.ts
@@ -1,19 +1,17 @@
 import { VisionCameraProxy } from 'react-native-vision-camera';
-import type {
-  Frame,
-  ScannerOptions,
-  BarcodeScannerPlugin,
-  Barcode,
-} from './types';
+import type { Frame, ScannerOptions, ScannerPlugin, Barcode } from './types';
 
 const LINKING_ERROR = `Can't load plugin scanBarcodes.Try cleaning cache or reinstall plugin.`;
 
 export function createBarcodeScannerPlugin(
   options?: ScannerOptions
-): BarcodeScannerPlugin {
-  const optionsSafe = options || {} as ScannerOptions;
+): ScannerPlugin {
+  const optionsSafe = options || ({} as ScannerOptions);
 
-  const plugin = VisionCameraProxy.initFrameProcessorPlugin('scanBarcodes', optionsSafe);
+  const plugin = VisionCameraProxy.initFrameProcessorPlugin(
+    'scanBarcodes',
+    optionsSafe
+  );
   if (!plugin) {
     throw new Error(LINKING_ERROR);
   }
